[ansatz] Moved the classes.
Todo:
- Check if we can make the pulses module cleaner.

Ideas:
- Rename PulseFactory and child classes.
- Switch from sampling method to magic function __call__().
- Remove PulseFactory completely (maybe this causes more boilerplate code).
