+++ pulse_opt/integrals/loss_functions.py
@@ -6,7 +6,10 @@ import numpy as np
 from quantum_gates.integrators import Integrator
 
 from .utilities import integrands
-from ..pulses.basis import PulseFactory, PowerFactory, FourierFactory, GaussianFactory
+from ..pulses.pulse_factory import PulseFactory
+from ..pulses.power_pulses import PowerFactory
+from ..pulses.fourier_pulses import FourierFactory
+from ..pulses.gaussian_pulses import GaussianFactory
 
 
 class Loss(object):
@@ -79,12 +82,12 @@ class PowerLoss(Loss):
     Example:
         .. code:: python
 
-import numpy as np
-from pulse_opt.integrals.loss_functions import PowerLoss
+            import numpy as np
+            from pulse_opt.integrals.loss_functions import PowerLoss
 
-loss = PowerLoss(shift=0.5, n=2)
-coeff = [1.0, 0.0, 0.0]  # n+1 coefficients
-print(f"The sum of all Ito integrals for a FourierPulse with coefficients {coeff} is {loss(coeff)}.")
+            loss = PowerLoss(shift=0.5, n=2)
+            coeff = [1.0, 0.0, 0.0]  # n+1 coefficients
+            print(f"The sum of all Ito integrals for a FourierPulse with coefficients {coeff} is {loss(coeff)}.")
     """
 
     def __init__(self, shift: float=0.5, n: int=3, weights: np.array=np.ones(8), theta: float=np.pi/2, a: float=1.0):
@@ -123,6 +126,37 @@ class FourierLoss(Loss):
         super(FourierLoss, self).__init__(
             factoryClass=FourierFactory,
             factoryArgs={"shift": shift, "n": n, "perform_checks": False},
+            weights=weights,
+            theta=theta,
+            a=a
+        )
+
+
+class GaussianLoss(Loss):
+    """ Computes the total variance of the Ito integrals for pulse based on Gaussians.
+
+    Args:
+        scale (float): Standard deviation of the Gaussians.
+        n (int): Number of Gaussian basis functions.
+        weights (np.array): Weight we give each Ito integral in the loss, defaults to np.ones(8).
+        theta (float): Upper limit of the integration, total area of the pulse, defaults to np.pi/2.
+        a (float): Scaling parameter in the Ito integrals, defaults to 1.0
+
+    Example:
+        .. code:: python
+
+            import numpy as np
+            from pulse_opt.integrals.loss_functions import GaussianLoss
+
+            loss = GaussianLoss(scale=0.25, n=4)
+            coeff = [0.5, 0.5, 0.5, 0.5]  # n coefficients
+            print(f"The sum of all Ito integrals for a GaussianPulse with coefficients {coeff} is {loss(coeff)}.")
+    """
+
+    def __init__(self, scale: float=0.25, n: int=3, weights: np.array=np.ones(8), theta: float=np.pi/2, a: float=1.0):
+        super(GaussianLoss, self).__init__(
+            factoryClass=GaussianFactory,
+            factoryArgs={"scale": scale, "n": n, "perform_checks": False},
             weights=weights,
             theta=theta,
             a=a
