+++ pulse_opt/pulses/basis.py
@@ -4,9 +4,6 @@
 
 import numpy as np
 from functools import cached_property
-from scipy.stats import norm
-
-from quantum_gates import Pulse
 
 
 class Basis(object):
@@ -137,210 +134,3 @@ class Basis(object):
         raise Exception(
             f"Could not generate default coefficient, as all basis functions have vanishing areas: {self.areas}."
         )
-
-
-class PulseFactory(object):
-    """ Takes a basis and provides a method to sample pulses with waveforms generated by this basis.
-
-    Args:
-        basis (Basis): Object representing the basis functions for the pulse.
-        perform_check (bool): Whether the constructed pulse should be verified.
-    """
-
-    def __init__(self, basis: Basis, perform_checks: bool=True):
-        self.basis = basis
-        self.perform_checks = perform_checks
-
-    def sample(self, coefficients):
-        """ Construct a pulse given normalized coefficients.
-
-        Note:
-            The user is responsible for giving coefficients such that the pulse is normalized. In scipy.optimize.minimize
-                this is possible via the constraint that is accessible in self.basis.constraints.
-        """
-        self._verify_coefficients(coefficients)
-        return Pulse(
-            pulse=self._get_waveform(coefficients),
-            parametrization=self._get_parametrization(coefficients),
-            perform_checks=self.perform_checks
-        )
-
-    def _get_waveform(self, coefficients) -> callable:
-        return self.basis.waveform(coefficients=coefficients)
-
-    def _get_parametrization(self, coefficients) -> callable:
-        return self.basis.parametrization(coefficients=coefficients)
-
-    def _verify_coefficients(self, coefficients):
-        """ Raises an exception if the coefficients are not valid.
-        """
-        if not self.basis.coefficient_are_valid(coefficients):
-            raise ValueError(f"Expected valid coefficients, but found otherwise. Coefficients: {coefficients}.")
-
-
-class PowerFactory(PulseFactory):
-    """Constructs pulses based on the power series.
-
-    Args:
-        n (int): Degree of the polynomial to be used.
-        shift (float): Number by which the basis functions should be shifted.
-    """
-
-    def __init__(self, shift: float=0.5, n: int=3, perform_checks=True):
-        self.shift = shift
-        self.n = n
-        super(PowerFactory, self).__init__(
-            basis=Basis(
-                functions=[(lambda x, power=i: x ** power) for i in range(n+1)],
-                integrals=[(lambda x, power=i: (x ** (power + 1))/(power + 1)) for i in range(n+1)],
-                shift=self.shift,
-                bounds=PowerFactory.get_bounds(n)
-            ),
-            perform_checks=perform_checks)
-
-    @staticmethod
-    def get_functions(n: int) -> list[callable]:
-        """ Generates a list of sin and cos functions that form the basis.
-
-        Args:
-            n (int): Degree of the polynomial to be used.
-
-        Returns:
-            Basis functions, polynomials up to degree n.
-        """
-
-        return [(lambda x, power=i: x ** power) for i in range(n+1)]
-
-    @staticmethod
-    def get_integrals(n: int) -> list[callable]:
-        """Generates a list of antiderivatives of the functions returned by the get_functions() method.
-
-        Args:
-            n (int): Degree of the polynomial to be used.
-
-        Returns:
-            Integrals of the basis functions.
-        """
-        return [(lambda x, power=i: (x ** (power + 1))/(power + 1)) for i in range(n+1)]
-
-    @staticmethod
-    def get_bounds(n: int) -> list[tuple]:
-        """ Generates the bounds on the function coeffcients. """
-        return [(1e-3, None)] + [(None, None) for i in range(n)]
-
-
-class FourierFactory(PulseFactory):
-    """Constructs pulses based on the Fourier series, namely sin (odd) and cos (even) functions.
-
-    Args:
-        n (int): Maximum number of zero crossings -> Creates 2(n + 1) basis functions.
-        shift (float): Number by which the basis functions should be shifted.
-    """
-
-    def __init__(self, shift: float=0.5, n: int=3, perform_checks=True):
-        self.shift = shift
-        self.n = n
-        super(FourierFactory, self).__init__(
-            basis=Basis(
-                functions=FourierFactory.get_functions(n),
-                integrals=FourierFactory.get_integrals(n),
-                shift=self.shift,
-                bounds=FourierFactory.get_bounds(n)
-            ),
-            perform_checks=perform_checks)
-
-    @staticmethod
-    def get_functions(n: int) -> list[callable]:
-        """ Generates a list of sin and cos functions that form the basis.
-
-        Args:
-            n (int): Maximum number of zero crossing in the basis functions.
-
-        Returns:
-            Basis functions based on sin() and cos() with increasing number of zero crossings. The first few functions
-                are up to constants of the form
-                [c*cos(c*x), c*sin(c*x), c*cos(2*c*x), c*sin(2*c*x), c*cos(3*c*x),...]
-                In this case, c = pi/2 such that the first function pair crosses zero at the boundary, and the i-th
-                function pair has i zero crossing, where i is an index (starts at 0).
-        """
-        return [(
-            lambda x, j=i: np.cos(x * (j//2+1) * (np.pi/2)) * (np.pi/2) if j % 2 == 0
-            else np.sin(x * (j//2+1) * (np.pi/2)) * (np.pi/2)
-        ) for i in range(2*(n+1))]
-
-    @staticmethod
-    def get_integrals(n: int) -> list[callable]:
-        """Generates a list of antiderivatives of the functions returned by the get_functions() method.
-
-        Args:
-            n (int): Maximum number of zero crossing in the basis functions.
-
-        Returns:
-            Integrals of the basis functions.
-        """
-        return [(
-            lambda x, j=i: np.sin(x * (j//2+1) * (np.pi/2)) - 0.0 if j % 2 == 0
-            else -np.cos(x * (j//2+1) * np.pi/2) + 1.0
-        ) for i in range(2*(n+1))]
-
-    @staticmethod
-    def get_bounds(n):
-        """ Generates the bounds on the function coeffcients. """
-        return [(1e-3, None), (1e-3, None)] + [(None, None) for i in range(2 * n)]
-
-
-class GaussianFactory(PulseFactory):
-    """Constructs pulses based on Gaussian pulses at fixed locations.
-
-    Args:
-        n (int): Number of Gaussians to be used.
-        scale (float): Standard deviations of the Gaussians.
-        perform_checks (bool): Should the resulting pulse be verified.
-    """
-
-    def __init__(self, n: int=10, scale: float=0.3, perform_checks=True):
-        self.n = n
-        self.scale = scale
-        super(GaussianFactory, self).__init__(
-            basis=Basis(
-                functions=GaussianFactory.get_functions(n=n, scale=scale),
-                integrals=GaussianFactory.get_integrals(n=n, scale=scale),
-                shift=0.0,
-                bounds=GaussianFactory.get_bounds(n=n)
-            ),
-            perform_checks=perform_checks)
-
-    @staticmethod
-    def get_functions(n: int, scale: float) -> list[callable]:
-        """Generates a list of n Gaussian functions with the same scale evenly distributed across [0,1].
-
-        Args:
-            n (int): Maximum number of zero crossing in the basis functions.
-            scale (float): Standard deviation of the Gaussians.
-
-        Returns:
-            List of Gaussians.
-        """
-        locations = np.linspace(0.0, 1.0, n) if n > 1 else [0.5]
-        return [lambda x, loc=location: norm.pdf(x, loc=loc, scale=scale) for location in locations]
-
-    @staticmethod
-    def get_integrals(n: int, scale: float) -> list[callable]:
-        """Generates a list of antiderivatives of the functions returned by the get_functions() method.
-
-        Args:
-            n (int): Maximum number of zero crossing in the basis functions.
-            scale (float): Standard deviation of the Gaussians.
-
-        Returns:
-            Integrals of the basis functions.
-        """
-        locations = np.linspace(0.0, 1.0, n) if n > 1 else [0.5]
-        return [
-            lambda x, loc=location: norm.cdf(x, loc=loc, scale=scale) - norm.cdf(0.0, loc=loc, scale=scale)
-            for location in locations
-        ]
-
-    @staticmethod
-    def get_bounds(n: int):
-        return [(None, None) for i in range(n)]
